+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support Draw.io Desktop v22.1.16
   - --svg-theme is now supported for SVG format
   - --svg-links-target is available for SVG format
+- Support '--all-pages' for exporting all pages into one PDF per drawio file
 
 ## [1.3.2] - 2025-06-04